# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BizCode API
  version: 0.1.0
  description: |
    REST API for BizCode commercial management system.
    Runs as an Express 5 sidecar on localhost:3001.
    Responses use a JSON envelope `{ success: true, data: ... }` except `/api/health`.
    Session authentication is required for protected resources.
    Protected endpoints can optionally receive `x-bizcode-channel` to enforce channel scope authorization.
    All routes accept optional `X-Request-Id` (max 128 characters); invalid or empty values are ignored and a UUID is assigned. The effective id is echoed on every response (`X-Request-Id`; see runtime middleware `correlationId`).
    Interactive documentation (Swagger UI): http://localhost:3001/api-docs/ (same contract as this file).
  contact:
    name: BizCode Team
servers:
  - url: http://localhost:3001
    description: Local development / Tauri sidecar
...
paths:
  /api/auth/setup-owner:
    post:
      tags: [auth]
      summary: Bootstrap first owner account
      description: One-time endpoint to create initial tenant and owner user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupOwnerInput'
      responses:
        '201':
          description: Owner created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupOwnerEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/auth/login:
    post:
      tags: [auth]
      summary: Create authenticated session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              example: 5
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '429':
          description: Account locked after too many failed login attempts
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              example: 5
            X-RateLimit-Remaining:
              schema:
                type: integer
              example: 0
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /api/auth/me:
    get:
      tags: [auth]
      summary: Get authenticated claims
      responses:
        '200':
          description: Current authenticated claims
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthClaimsEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: bizcode_session
...
  schemas:
...
    ApiErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
    SetupOwnerInput:
      type: object
      required: [tenantName, tenantSlug, username, password]
      properties:
        tenantName:
          type: string
          minLength: 2
        tenantSlug:
          type: string
          minLength: 2
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
    SetupOwnerResult:
      type: object
      required: [tenantId, userId, role]
      properties:
        tenantId:
          type: integer
        userId:
          type: integer
        role:
          type: string
    SetupOwnerEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/SetupOwnerResult'
    LoginInput:
      type: object
      required: [tenantSlug, username, password]
      properties:
        tenantSlug:
          type: string
        username:
          type: string
        password:
          type: string
    LoginResult:
      type: object
      required: [userId, tenantId, username, role]
      properties:
        userId:
          type: integer
        tenantId:
          type: integer
        username:
          type: string
        role:
          type: string
    LoginEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/LoginResult'
...
    AuthScope:
      type: object
      required: [tenantId, branchIds, warehouseIds, routeIds, channels]
      properties:
        tenantId:
          type: integer
        branchIds:
          type: array
          items:
            type: integer
        warehouseIds:
          type: array
          items:
            type: integer
        routeIds:
          type: array
          items:
            type: integer
        channels:
          type: array
          items:
            type: string
    AuthClaims:
      type: object
      required: [userId, tenantId, username, role, permissions, scope]
      properties:
        userId:
          type: integer
        tenantId:
          type: integer
        username:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        scope:
          $ref: '#/components/schemas/AuthScope'
    AuthClaimsEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/AuthClaims'
...
  responses:
    BadRequest:
      description: Request payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
    UnauthorizedError:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
...
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
...